move GMCommand and PlayerCommand to their own modules
