+++ arptypes/src/lib.rs
@@ -1,4 +1,10 @@
+pub mod command;
+pub mod log;
 pub mod types;
+
+pub use command::*;
+pub use log::*;
+pub use types::*;
 
 // I don't know why I still need to use old-style extern crate for uom.
 // In types.rs we use "ISQ!", but replacing this `extern crate` with
