+++ arptypes/src/types.rs
@@ -11,6 +11,8 @@ use ts_rs::TS;
 use uom::si::length::{centimeter, meter};
 use uuid::{Error as UuidParseError, Uuid};
 
+use crate::{GMCommand, GameLog};
+
 pub type Color = String;
 pub type Inventory = HashMap<ItemID, u64>;
 pub type Terrain = Vec<Point3>;
@@ -432,355 +434,6 @@ pub enum Dice {
   BestOf(u8, Box<Dice>),
 }
 
-/// Top-level commands that can be sent from a Player
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize, TS)]
-#[serde(tag = "t")]
-pub enum PlayerCommand {
-  // should these enums include PlayerID? probably not...
-  ChatFromPlayer {
-    message: String,
-  },
-
-  CreateNote {
-    path: FolderPath,
-    note: Note,
-  },
-
-  EditNote {
-    // This FolderPath is scoped to the player's folder.
-    path: FolderPath,
-    original_name: String,
-    note: Note,
-  },
-
-  // Out-of-combat actions:
-  // /// Use an Ability out of combat.
-  // ActCreature { // NYI
-  //   creature_id: CreatureID,
-  //   ability_id: AbilityID,
-  //   target: DecidedTarget,
-  // },
-  /// Move a creature along a path within a scene.
-  /// There must be a clear path according to the current loaded map. It doesn't matter whether
-  /// the creature is in combat.
-  PathCreature {
-    creature_id: CreatureID,
-    destination: Point3,
-  },
-
-  // In-combat actions:
-  /// Make the current creature use an ability.
-  CombatAct {
-    ability_id: AbilityID,
-    target: DecidedTarget,
-  },
-  /// Move the current creature in combat to a point.
-  /// There must be a clear path according to the current loaded map.
-  PathCurrentCombatCreature {
-    destination: Point3,
-  },
-  /// End the current creature's turn.
-  EndTurn,
-  // GiveItem or something should be here
-}
-
-/// Top-level commands that can be sent from a GM to affect the state of the game.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, TS)]
-#[serde(tag = "t")]
-pub enum GMCommand {
-  LoadModule {
-    name: String,
-    source: ModuleSource,
-    game: Game,
-    path: FolderPath,
-  },
-
-  ChatFromGM {
-    message: String,
-  },
-
-  AttributeCheck {
-    creature_id: CreatureID,
-    attribute_check: AttributeCheck,
-  },
-
-  /// Create a folder, given segments leading to it.
-  CreateFolder {
-    path: FolderPath,
-  },
-  /// Rename a folder. DEPRECATED (I think?)
-  RenameFolder {
-    path: FolderPath,
-    new_name: String,
-  },
-
-  /// Move some object from one folder to another.
-  MoveFolderItem {
-    source: FolderPath,
-    item_id: FolderItemID,
-    destination: FolderPath,
-  },
-  /// Copy an object to a folder. It's okay to copy it to the same folder.
-  CopyFolderItem {
-    source: FolderPath,
-    item_id: FolderItemID,
-    dest: FolderPath,
-  },
-  DeleteFolderItem {
-    path: FolderPath,
-    item_id: FolderItemID,
-  },
-  RenameFolderItem {
-    path: FolderPath,
-    item_id: FolderItemID,
-    new_name: String,
-  },
-
-  /// Create an Item in a folder. (this will probably take an ItemCreation in the future)
-  CreateItem {
-    path: FolderPath,
-    name: String,
-  },
-  /// Edit an Item. The ID in the given Item must match an existing Item.
-  EditItem {
-    item: Item,
-  },
-
-  /// Create a Note inside of a Folder.
-  CreateNote {
-    path: FolderPath,
-    note: Note,
-  },
-  /// Edit a Note inside of a Folder.
-  EditNote {
-    path: FolderPath,
-    original_name: String,
-    note: Note,
-  },
-
-  // ** Inventory management **
-  // These work for creatures or scenes
-  TransferItem {
-    from: InventoryOwner,
-    to: InventoryOwner,
-    item_id: ItemID,
-    count: u64,
-  },
-  RemoveItem {
-    owner: InventoryOwner,
-    item_id: ItemID,
-    count: u64,
-  },
-  SetItemCount {
-    owner: InventoryOwner,
-    item_id: ItemID,
-    count: u64,
-  },
-
-  // ** Scene management **
-  /// Create a Scene.
-  CreateScene {
-    path: FolderPath,
-    scene: SceneCreation,
-  },
-  EditSceneDetails {
-    scene_id: SceneID,
-    details: SceneCreation,
-  },
-  SetSceneCreatureVisibility {
-    scene_id: SceneID,
-    creature_id: CreatureID,
-    visibility: Visibility,
-  },
-  AddCreatureToScene {
-    scene_id: SceneID,
-    creature_id: CreatureID,
-    visibility: Visibility,
-  },
-  RemoveCreatureFromScene {
-    scene_id: SceneID,
-    creature_id: CreatureID,
-  },
-  AddSceneChallenge {
-    scene_id: SceneID,
-    description: String,
-    challenge: AttributeCheck,
-  },
-  RemoveSceneChallenge {
-    scene_id: SceneID,
-    description: String,
-  },
-  SetFocusedSceneCreatures {
-    scene_id: SceneID,
-    creatures: Vec<CreatureID>,
-  },
-  // AddSceneVolumeCondition {
-  //   scene_id: SceneID,
-  //   point: Point3,
-  //   volume: Volume,
-  //   condition: Condition,
-  //   duration: Duration,
-  // },
-  RemoveSceneVolumeCondition {
-    scene_id: SceneID,
-    condition_id: ConditionID,
-  },
-  EditSceneTerrain {
-    scene_id: SceneID,
-    #[ts(type = "Terrain")]
-    terrain: Vec<Point3>,
-  },
-  EditSceneHighlights {
-    scene_id: SceneID,
-    #[ts(type = "Highlights")]
-    highlights: HashMap<Point3, (Color, Visibility)>,
-  },
-  EditSceneAnnotations {
-    scene_id: SceneID,
-    #[ts(type = "Annotations")]
-    annotations: HashMap<Point3, (String, Visibility)>,
-  },
-  EditSceneRelatedScenes {
-    scene_id: SceneID,
-    #[ts(type = "RelatedScenes")]
-    related_scenes: HashSet<SceneID>,
-  },
-  EditSceneSceneHotspots {
-    scene_id: SceneID,
-    #[ts(type = "SceneHotspots")]
-    scene_hotspots: HashMap<Point3, SceneID>,
-  },
-
-  // ** Combat management **
-  /// Start a combat with the specified creatures.
-  StartCombat {
-    scene_id: SceneID,
-    combatants: Vec<CreatureID>,
-  },
-  /// Stop the current combat.
-  StopCombat,
-  /// Add a creature to combat.
-  AddCreatureToCombat {
-    creature_id: CreatureID,
-  },
-  /// Remove a creature from combat.
-  RemoveCreatureFromCombat {
-    creature_id: CreatureID,
-  },
-  /// Modify a creature's order in the combat list.
-  ChangeCreatureInitiative {
-    creature_id: CreatureID,
-    initiative: i16,
-  },
-  /// Reroll initiative for all creatures in combat, and sort the combat list
-  RerollCombatInitiative,
-  /// Move to the next creature in the initiative list. This does *not* run any end-of-turn or
-  /// start-turn events.
-  ForceNextTurn,
-  /// Move to the previous creature in the initiative list. This does *not* run any end-of-turn or
-  /// start-turn events.
-  ForcePrevTurn,
-
-  // ** Combat **
-  /// Use an Ability out of combat.
-  ActCreature {
-    scene_id: SceneID,
-    creature_id: CreatureID,
-    ability_id: AbilityID,
-    target: DecidedTarget,
-  },
-  /// Make the current creature use an ability.
-  CombatAct {
-    ability_id: AbilityID,
-    target: DecidedTarget,
-  },
-  /// Move the current creature in combat to a point.
-  /// There must be a clear path according to the current loaded map.
-  PathCurrentCombatCreature {
-    destination: Point3,
-  },
-  /// End the current creature's turn.
-  EndTurn,
-
-  // ** Classes & Abilities **
-  CreateClass {
-    path: FolderPath,
-    class: ClassCreation,
-  },
-  EditClass {
-    class: Class,
-  },
-  CreateAbility {
-    path: FolderPath,
-    ability: AbilityCreation,
-  },
-  EditAbility {
-    ability: Ability,
-  },
-
-  // ** Creature Manipulation **
-  /// Create a new creature.
-  CreateCreature {
-    path: FolderPath,
-    creature: CreatureCreation,
-  },
-  /// Edit an existing creature.
-  EditCreatureDetails {
-    creature: Creature,
-  },
-  /// Assign a creature's position within a scene.
-  SetCreaturePos {
-    scene_id: SceneID,
-    creature_id: CreatureID,
-    pos: Point3,
-  },
-  /// Move a creature along a path within a scene.
-  /// There must be a clear path according to the current loaded map. It doesn't matter whether
-  /// the creature is in combat.
-  PathCreature {
-    scene_id: SceneID,
-    creature_id: CreatureID,
-    destination: Point3,
-  },
-
-  // ** Player Manipulation **
-  /// Register a player as available for controlling a creature.
-  RegisterPlayer {
-    id: PlayerID,
-  },
-  /// Give control of a creature to a player.
-  GiveCreaturesToPlayer {
-    player_id: PlayerID,
-    creature_ids: Vec<CreatureID>,
-  },
-  /// Remove a player from the game, allowing all of their creatures to be given to other players.
-  UnregisterPlayer {
-    id: PlayerID,
-  },
-  /// Remove control of a creature from a player.
-  RemoveCreaturesFromPlayer {
-    player_id: PlayerID,
-    creature_ids: Vec<CreatureID>,
-  },
-  /// Move a player to a particular scene, so they only see what's happening in that scene.
-  /// Note that this doesn't have any affect on a player's *characters*.
-  /// The scene name can be None (null) to not show any scene to the player.
-  SetPlayerScene {
-    player_id: PlayerID,
-    scene_id: Option<SceneID>,
-  },
-
-  SetActiveScene {
-    id: Option<SceneID>,
-  },
-
-  /// Roll back to a specific snapshot + log index
-  Rollback {
-    snapshot_index: usize,
-    log_index: usize,
-  },
-}
-
 /// Potential targets for an ability.
 #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize, TS)]
 pub enum PotentialTargets {
@@ -1191,313 +844,4 @@ impl Folder {
 pub struct ChangedGame {
   pub game: Game,
   pub logs: Vec<GameLog>,
-}
-
-/// A representation of state change in a Creature. See `GameLog`.
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize, TS)]
-#[serde(tag = "t")]
-pub enum CreatureLog {
-  Damage { hp: HP, rolls: Vec<i16> },
-  Heal { hp: HP, rolls: Vec<i16> },
-  GenerateEnergy { energy: Energy },
-  ReduceEnergy { energy: Energy },
-  ApplyCondition { id: ConditionID, duration: Duration, condition: Condition },
-  DecrementConditionRemaining { id: ConditionID },
-  RemoveCondition { id: ConditionID },
-}
-
-// TODO: get rid of CombatLog, it's dumb... unless we ever support multiple Combats?
-/// Representation of state changes in a Combat. See `GameLog`.
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize, TS)]
-#[serde(tag = "t")]
-pub enum CombatLog {
-  // Consume some of the movement from the current combat-creatur
-  ConsumeMovement {
-    #[ts(type = "number")]
-    distance: u32units::Length,
-  },
-  ChangeCreatureInitiative {
-    creature_id: CreatureID,
-    initiative: i16,
-  },
-  EndTurn {
-    creature_id: CreatureID,
-  }, // the end of this creature's turn
-  ForceNextTurn,
-  ForcePrevTurn,
-  RerollInitiative {
-    combatants: Vec<(CreatureID, i16)>,
-  },
-}
-
-pub fn creature_logs_into_game_logs(creature_id: CreatureID, ls: Vec<CreatureLog>) -> Vec<GameLog> {
-  ls.into_iter().map(|log| GameLog::CreatureLog { creature_id, log }).collect()
-}
-
-/// Representation of a change to the game state. All change to the game happens via these values.
-/// Note that these represent *concrete* changes to the game, which will have deterministic results.
-/// i.e., randomness happens when processing `GMCommand`s, which then result in specific
-/// `GameLog`s.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, TS)]
-#[serde(tag = "t")]
-pub enum GameLog {
-  LoadModule {
-    name: String,
-    source: ModuleSource,
-    #[ts(skip)]
-    module: Game,
-    path: FolderPath,
-  },
-
-  SetActiveScene {
-    id: Option<SceneID>,
-  },
-
-  // ** Player Manipulation **
-  RegisterPlayer {
-    id: PlayerID,
-  },
-  GiveCreaturesToPlayer {
-    player_id: PlayerID,
-    creature_ids: Vec<CreatureID>,
-  },
-  UnregisterPlayer {
-    id: PlayerID,
-  },
-  RemoveCreaturesFromPlayer {
-    player_id: PlayerID,
-    creature_ids: Vec<CreatureID>,
-  },
-  SetPlayerScene {
-    player_id: PlayerID,
-    scene_id: Option<SceneID>,
-  },
-
-  ChatFromGM {
-    message: String,
-  },
-  ChatFromPlayer {
-    player_id: PlayerID,
-    message: String,
-  },
-
-  AttributeCheckResult {
-    creature_id: CreatureID,
-    attribute_check: AttributeCheck,
-    actual: u8,
-    success: bool,
-  },
-
-  // ** Folder Management **
-  /// Create a folder, given segments leading to it.
-  CreateFolder {
-    path: FolderPath,
-  },
-  /// Rename a folder. DEPRECATED (I think?)
-  RenameFolder {
-    path: FolderPath,
-    new_name: String,
-  },
-  MoveFolderItem {
-    source: FolderPath,
-    item_id: FolderItemID,
-    destination: FolderPath,
-  },
-  CopyFolderItem {
-    source: FolderPath,
-    item_id: FolderItemID,
-    dest: FolderPath,
-    new_item_id: FolderItemID,
-  },
-  DeleteFolderItem {
-    path: FolderPath,
-    item_id: FolderItemID,
-  },
-  RenameFolderItem {
-    path: FolderPath,
-    item_id: FolderItemID,
-    new_name: String,
-  },
-
-  CreateItem {
-    path: FolderPath,
-    item: Item,
-  },
-  EditItem {
-    item: Item,
-  },
-
-  CreateNote {
-    path: FolderPath,
-    note: Note,
-  },
-  EditNote {
-    path: FolderPath,
-    original_name: String,
-    note: Note,
-  },
-
-  // ** Inventory management **
-  TransferItem {
-    from: InventoryOwner,
-    to: InventoryOwner,
-    item_id: ItemID,
-    count: u64,
-  },
-  RemoveItem {
-    owner: InventoryOwner,
-    item_id: ItemID,
-    count: u64,
-  },
-  SetItemCount {
-    owner: InventoryOwner,
-    item_id: ItemID,
-    count: u64,
-  },
-
-  CreateScene {
-    path: FolderPath,
-    scene: Scene,
-  },
-  EditSceneDetails {
-    scene_id: SceneID,
-    details: SceneCreation,
-  },
-  SetSceneCreatureVisibility {
-    scene_id: SceneID,
-    creature_id: CreatureID,
-    visibility: Visibility,
-  },
-  AddCreatureToScene {
-    scene_id: SceneID,
-    creature_id: CreatureID,
-    visibility: Visibility,
-  },
-  RemoveCreatureFromScene {
-    scene_id: SceneID,
-    creature_id: CreatureID,
-  },
-  AddSceneChallenge {
-    scene_id: SceneID,
-    description: String,
-    challenge: AttributeCheck,
-  },
-  RemoveSceneChallenge {
-    scene_id: SceneID,
-    description: String,
-  },
-  SetFocusedSceneCreatures {
-    scene_id: SceneID,
-    creatures: Vec<CreatureID>,
-  },
-  RemoveSceneVolumeCondition {
-    scene_id: SceneID,
-    condition_id: ConditionID,
-  },
-
-  EditSceneTerrain {
-    scene_id: SceneID,
-    #[ts(type = "Terrain")]
-    terrain: Vec<Point3>,
-  },
-  EditSceneHighlights {
-    scene_id: SceneID,
-    #[ts(type = "Highlights")]
-    highlights: HashMap<Point3, (Color, Visibility)>,
-  },
-  EditSceneAnnotations {
-    scene_id: SceneID,
-    #[ts(type = "Annotations")]
-    annotations: HashMap<Point3, (String, Visibility)>,
-  },
-  EditSceneRelatedScenes {
-    scene_id: SceneID,
-    #[ts(type = "RelatedScenes")]
-    related_scenes: HashSet<SceneID>,
-  },
-  EditSceneSceneHotspots {
-    scene_id: SceneID,
-    #[ts(type = "SceneHotspots")]
-    scene_hotspots: HashMap<Point3, SceneID>,
-  },
-
-  CombatLog {
-    log: CombatLog,
-  },
-  /// A creature log wrapped in a game log.
-  CreatureLog {
-    creature_id: CreatureID,
-    log: CreatureLog,
-  },
-  SetCreaturePos {
-    scene_id: SceneID,
-    creature_id: CreatureID,
-    pos: Point3,
-  },
-  PathCreature {
-    scene_id: SceneID,
-    creature_id: CreatureID,
-    path: Vec<Point3>,
-  },
-
-  AddVolumeCondition {
-    scene_id: SceneID,
-    point: Point3,
-    volume: Volume,
-    condition_id: ConditionID,
-    condition: Condition,
-    duration: Duration,
-  },
-
-  StartCombat {
-    scene_id: SceneID,
-    combatants: Vec<(CreatureID, i16)>,
-  },
-  StopCombat,
-
-  // ** Classes & Abilities **
-  CreateClass {
-    path: FolderPath,
-    class: Class,
-  },
-  EditClass {
-    class: Class,
-  },
-  CreateAbility {
-    path: FolderPath,
-    ability: Ability,
-  },
-  EditAbility {
-    ability: Ability,
-  },
-
-  // ** Creatures **
-  CreateCreature {
-    path: FolderPath,
-    creature: Creature,
-  },
-  // Deprecated: please migrate & remove
-  EditCreatureDetails {
-    creature_id: CreatureID,
-    details: CreatureCreation,
-  },
-  EditCreature {
-    creature: Creature,
-  },
-  AddCreatureToCombat {
-    creature_id: CreatureID,
-    initiative: i16,
-  },
-  RemoveCreatureFromCombat {
-    creature_id: CreatureID,
-  },
-  Rollback {
-    // This is purely informational?
-    snapshot_index: usize,
-    log_index: usize,
-  },
-}
-
-pub fn combat_logs_into_game_logs(ls: Vec<CombatLog>) -> Vec<GameLog> {
-  ls.into_iter().map(|log| GameLog::CombatLog { log }).collect()
 }
